Securing store management views

+++ store_management/views.py
@@ -2,10 +2,16 @@ from django.shortcuts import render, reverse, redirect, get_object_or_404
 from .forms import ProductForm
 from products.models import Product, Category
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 
+@login_required
 def add_product(request):
     """ Add's products to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners are\
+             allowed on this page!')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -23,8 +29,13 @@ def add_product(request):
     return render(request, 'store_management/add_product.html', context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit product's in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners are\
+             allowed on this page!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -46,8 +57,13 @@ def edit_product(request, product_id):
     return render(request, 'store_management/edit_product.html', context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete product's from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners are\
+             allowed on this page!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.info(request, 'Product deleted successfully.')
